implemented Aruns review for find word in sentence program
